o Work in progress
When merging two .mbtiles, if the destination slot for the tile
is already occupied, blend the source image instead of replacing.

TODO: Optimize by selecting the intersections and storing it on a
list, saving a sql query for every single tile.

TODO: Make it a command line option for the --merge group.